get static file

+++ edc_utils/get_static_file.py
@@ -6,19 +6,21 @@ from urllib.error import URLError
 
 
 def get_static_file(app_label, filename):
-    path = os.path.join(settings.STATIC_ROOT, app_label, filename)
-    try:
-        with open(path, "r"):
-            pass
-    except FileNotFoundError:
-        path = os.path.join(f"https://{settings.STATIC_URL}", app_label, filename)
+    path = os.path.join(settings.STATIC_ROOT or "", app_label, filename)
+    if os.path.isfile(path):
         try:
-            urllib.request.urlretrieve(path)
-        except URLError:
-            raise FileNotFoundError(
-                f"Static file not found. Tried "
-                f"STATIC_ROOT ({settings.STATIC_ROOT}) and "
-                f"STATIC_URL ({settings.STATIC_URL}). "
-                f"Got {app_label}/{filename}."
-            )
+            with open(path, "r"):
+                pass
+        except FileNotFoundError:
+            path = os.path.join(
+                f"https://{settings.STATIC_URL}", app_label, filename)
+            try:
+                urllib.request.urlretrieve(path)
+            except URLError:
+                raise FileNotFoundError(
+                    f"Static file not found. Tried "
+                    f"STATIC_ROOT ({settings.STATIC_ROOT}) and "
+                    f"STATIC_URL ({settings.STATIC_URL}). "
+                    f"Got {app_label}/{filename}."
+                )
     return path
